move household data-structure and pipeline to separate component
