+++ src/vivarium_census_prl_synth_pop/components/household_migration.py
@@ -1,15 +1,9 @@
-import numpy as np
 import pandas as pd
 from vivarium.framework.engine import Builder
 from vivarium.framework.event import Event
-from vivarium.framework.population import SimulantData
 from vivarium.framework.time import get_time_stamp
 
 from vivarium_census_prl_synth_pop.constants import metadata, paths
-from vivarium_census_prl_synth_pop.constants.data_values import (
-    GQ_HOUSING_TYPE_MAP,
-    HOUSING_TYPES,
-)
 from vivarium_census_prl_synth_pop.utilities import update_address_ids
 
 
@@ -39,8 +33,8 @@ class HouseholdMigration:
     #################
 
     def setup(self, builder: Builder):
-        self.config = builder.configuration
         self.start_time = get_time_stamp(builder.configuration.time.start)
+        self.households = builder.components.get_component("households")
 
         move_rate_data = builder.lookup.build_table(
             data=pd.read_csv(
@@ -56,24 +50,12 @@ class HouseholdMigration:
 
         self.randomness = builder.randomness.get_stream(self.name)
         self.columns_used = [
-            "tracked",
             "household_id",
             "relation_to_household_head",
         ]
 
         self.population_view = builder.population.get_view(self.columns_used)
 
-        self.households = None
-        self.household_details = builder.value.register_value_producer(
-            "household_details",
-            source=self.get_household_details,
-            requires_columns=["household_id", "tracked"],
-        )
-
-        builder.population.initializes_simulants(
-            self.on_initialize_simulants,
-            requires_columns=["household_id"],
-        )
         builder.event.register_listener(
             "time_step",
             self.on_time_step,
@@ -84,24 +66,12 @@ class HouseholdMigration:
     # Event-driven methods #
     ########################
 
-    def on_initialize_simulants(self, pop_data: SimulantData) -> None:
-        """
-        add addresses to each household in the population table
-        """
-        if pop_data.creation_time < self.start_time:  # initial pop
-            self.households = self.generate_initial_households(pop_data)
-
     def on_time_step(self, event: Event):
         """
         choose which households move;
         move those households to a new address
         """
-        # NOTE: Currently, it is possible for a household not to have a living reference person;
-        # in this case, that household can no longer move.
-        pop = self.population_view.get(
-            event.index,
-            query="alive == 'alive' and tracked",
-        )
+        pop = self.population_view.get(event.index, query="alive == 'alive'")
         reference_people = pop[pop["relation_to_household_head"] == "Reference person"]
 
         household_sizes = pop.groupby("household_id").size()
@@ -118,45 +88,11 @@ class HouseholdMigration:
             "moving_households",
         )
 
-        self.households, _ = update_address_ids(
-            moving_units=self.households,
+        self.households.households, _ = update_address_ids(
+            moving_units=self.households.households,
             units_that_move_ids=movers["household_id"],
-            starting_address_id=self.households["address_id"].max() + 1,
+            starting_address_id=self.households.households["address_id"].max() + 1,
             address_id_col_name="address_id",
         )
 
         self.population_view.update(pop)
-
-    ##################
-    # Helper methods #
-    ##################
-
-    def generate_initial_households(self, pop_data: SimulantData) -> pd.DataFrame():
-        households = self.population_view.subview(["household_id"]).get(pop_data.index)
-        households = (
-            households.drop_duplicates().sort_values("household_id").set_index("household_id")
-        )
-        households["address_id"] = np.arange(len(households))
-        households["housing_type"] = households.index.map(GQ_HOUSING_TYPE_MAP).fillna(
-            "Standard"
-        )
-        # set housing type dtype
-        households["housing_type"] = households["housing_type"].astype(
-            pd.CategoricalDtype(categories=HOUSING_TYPES)
-        )
-
-        return households
-
-    def get_household_details(self, idx: pd.Index) -> pd.DataFrame:
-        """Source of the household_details pipeline"""
-        pop = self.population_view.get(idx)[["household_id"]]
-        household_details = pop.join(
-            self.households,
-            on="household_id",
-        )
-        # FIXME: why is the `housing_type` column losing its CategoricalDtype?
-        household_details["housing_type"] = household_details["housing_type"].astype(
-            pd.CategoricalDtype(categories=HOUSING_TYPES)
-        )
-
-        return household_details
