+++ src/vivarium_census_prl_synth_pop/model_specifications/model_spec.yaml
@@ -2,6 +2,7 @@ components:
     vivarium_census_prl_synth_pop:
         components:
             - Population()
+            - Households()
             - HouseholdMigration()
             - PersonMigration()
             - HouseholdEmigration()
